Split RoutingTableWorker into WorkerRegistrationActor and RoutingTableUpdateActor.
RoutingTableUpdateActor performs two independent tasks: worker registration and
routing table update. Split RoutingTableWorker into two actors and make each actor
perform only one task.
