+++ app/beyond/route/RoutingTableConfig.scala
@@ -1,6 +1,5 @@
 package beyond.route
 
 object RoutingTableConfig {
-  val WorkersPath: String = "/workers"
   val RoutingTablePath: String = "/routing-table"
 }
